Login and Registar Pages Completed
Login and register pages are all completed, front and back end. All the features are correctly working. The only little problem that persists is the logout button being unable to clear the username inside the button correctly when we go back after logging out, although its redirect is actually correct since it sends us to the login.aspx with all the session variables cleared and abandoned.