Disponibilidade Formador Development
Did some development on the Disponibilidade part of the project. Already can show the calendar and click to create an event of "Não Disponível" either daily, weekly or monthly. Also have buttons to mark all saturdays, saturday mornings. saturday afternoons, weekdays mornings and weekdays afternoons as "não disponível". From now on I need to find a way to send the selectedslots array that is used to store all the selected slots into the server and then convert the data into a list so I can use it from the C# area and send it to the DB