FullCalendar Read DB and Insert
Made some work on reading the DB data and to insert it into the fullcalendar via events. Currently I'm only able to fetch each time_slot one by one and I need to group them up when they are contiguous so it is shown a little bit better on the calendar for the user. Currently it is all badly formatted because each day that has events on it has events with 1h each. Which makes it awful